Add toast middleware
